requirements.txt
Flask
PyMySQL
mysql-connector-python

